feat : inputType crud 생성

+++ src/main/kotlin/com/Dnight/calinify/common/inputType/InputTypeRepository.kt
@@ -2,4 +2,4 @@ package com.dnight.calinify.common.inputType
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface InputTypeRepository : JpaRepository<InputTypeEntity, Long>
+interface InputTypeRepository : JpaRepository<InputTypeEntity, Int>
